Create component CityListItem
